+++ packages/pitchset/index.js
@@ -8,7 +8,7 @@
  * @module pitchset
  */
 import { chroma } from 'tonal-note'
-import { map } from 'tonal-array'
+import { map, asArr } from 'tonal-array'
 
 function toInt (set) { return parseInt(toBinary(set), 2) }
 
@@ -38,6 +38,7 @@ export function toBinary (set) {
  * pitchset.equal('c2 d3', 'c5 d2') // => true
  */
 export function equal (s1, s2) {
+  if (arguments.length === 1) return function (s) { return equal(s1, s) }
   return toBinary(s1) === toBinary(s2)
 }
 
@@ -51,6 +52,7 @@ export function equal (s1, s2) {
  * pitchset.subset('c d e', 'C2 D4 D5 C6') // => true
  */
 export function subset (set, test) {
+  if (arguments.length === 1) return function (t) { return subset(set, t) }
   test = toInt(test)
   return (test & toInt(set)) === test
 }
@@ -64,6 +66,37 @@ export function subset (set, test) {
  * pitchset.subset('c d e', 'C2 D4 F4 D5 E5 C6') // => true
  */
 export function superset (set, test) {
+  if (arguments.length === 1) return function (t) { return superset(set, t) }
   test = toInt(test)
   return (test | toInt(set)) === test
+}
+
+/**
+ * Test if a given pitch set includes a note
+ * @param {Array|String} set - the base set to test against
+ * @param {String|Pitch} note - the note to test
+ * @return {Boolean} true if the note is included in the pitchset
+ * @example
+ * pitchset.includes('c d e', 'C4') // =A true
+ * pitchset.includes('c d e', 'C#4') // =A false
+ */
+export function includes (set, note) {
+  if (arguments.length > 1) return includes(set)(note)
+  set = toBinary(set)
+  return function (note) { return set[chroma(note)] === '1' }
+}
+
+/**
+ * Filter a list with a pitch set
+ *
+ * @param {Array|String} set - the pitch set
+ * @param {Array|String} notes - the note list to be filtered
+ * @return {Array} the filtered notes
+ *
+ * @example
+ * pitchset.filter('c d e', 'c2 c#2 d2 c3 c#3 d3') // => [ 'c2', 'd2', 'c3', 'd3' ])
+ */
+export function filter (set, notes) {
+  if (arguments.length === 1) return function (n) { return filter(set, n) }
+  return asArr(notes).filter(includes(set))
 }